Conflicts:
	driver/run_dense_fanboth_mpi.cpp
	src/ETree.cpp

+++ src/FBMatrix.cpp
@@ -38,7 +38,7 @@ namespace LIBCHOLESKY{
 
 
     //for 2D maps
-    prow = sqrt(np);
+    prow = (Int)sqrt((double)np);
 
 
 #ifdef MAP2D
